events with participants, not groups, also added get wishlist by user endpoint

+++ backend/src/main/java/ru/secretsanta/service/wishlist/WishlistServiceImpl.java
@@ -9,6 +9,7 @@ import ru.secretsanta.entity.user.User;
 import ru.secretsanta.entity.wishlist.WishlistItem;
 import ru.secretsanta.exception.common.NotFoundException;
 import ru.secretsanta.exception.common.TooMuchItemsException;
+import ru.secretsanta.exception.user.InsufficientPrivilegesException;
 import ru.secretsanta.mapper.wishlist.WishlistMapper;
 import ru.secretsanta.repository.UserRepository;
 import ru.secretsanta.repository.WishlistRepository;
@@ -25,8 +26,8 @@ public class WishlistServiceImpl implements WishlistService {
 
 
     @Override
-    public WishlistItemResponse addItem(String username, WishlistItem item) {
-        User user = userRepository.findByName(username)
+    public WishlistItemResponse addItem(Long userId, WishlistItem item) {
+        User user = userRepository.findById(userId)
                 .orElseThrow(() -> new NotFoundException("User not found"));
         item.setUser(user);
         List<WishlistItem> wishlistItemsOfUser = wishlistRepository.findByUser(user,Pageable.unpaged()).getContent();
@@ -37,9 +38,12 @@ public class WishlistServiceImpl implements WishlistService {
     }
 
     @Override
-    public WishlistItemResponse updateItem(String username, Long itemId, WishlistItem updatedItem) {
+    public WishlistItemResponse updateItem(Long userId, Long itemId, WishlistItem updatedItem) {
         WishlistItem existing = wishlistRepository.findById(itemId)
-                .orElseThrow(() -> new RuntimeException("Item not found"));
+                .orElseThrow(() -> new NotFoundException("Item not found"));
+        if (!existing.getUser().getId().equals(userId)){
+            throw new InsufficientPrivilegesException("you have no rights to change this item");
+        }
         existing.setTitle(updatedItem.getTitle());
         existing.setDescription(updatedItem.getDescription());
         existing.setPriority(updatedItem.getPriority());
@@ -47,14 +51,19 @@ public class WishlistServiceImpl implements WishlistService {
     }
 
     @Override
-    public void deleteItem(Long itemId) {
+    public void deleteItem(Long userId, Long itemId) {
+        WishlistItem existing = wishlistRepository.findById(itemId)
+                .orElseThrow(() -> new NotFoundException("Item not found"));
+        if (!existing.getUser().getId().equals(userId)){
+            throw new InsufficientPrivilegesException("you have no rights to delete this item");
+        }
         wishlistRepository.deleteById(itemId);
     }
 
     @Override
-    public Page<WishlistItemResponse> getWishlist(String username, Pageable pageable) {
-        User user = userRepository.findByName(username)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+    public Page<WishlistItemResponse> getWishlist(Long userId, Pageable pageable) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new NotFoundException("User not found"));
         return wishlistRepository.findByUser(user,pageable).map(WishlistMapper::toWishlistItemResponse);
     }
 }
